ТЗ-14, резерв 2

+++ src/main/java/ru/practicum/shareit/item/controller/ItemController.java
@@ -43,7 +43,7 @@ public class ItemController {
     }
 
     @GetMapping("/{itemId}")
-    public ItemDto getItemById(@PathVariable int itemId,
+    public ItemDto getItemById(@PathVariable long itemId,
                                @RequestHeader(USER_ID_HEADER) long userId) {
         return itemService.getItemById(itemId, userId);
     }
